Elevate `gitbutler-fs` and merge `gitbutler-storage` into it.
Clean it up and minimize it while at it.
