+++ crates/but-forge-storage/Cargo.toml
@@ -14,8 +14,9 @@ rust-version = "1.89"
 doctest = false
 
 [dependencies]
+but-secret.workspace = true
+but-fs = { workspace = true, features = ["legacy"] }
+
 serde.workspace = true
 anyhow.workspace = true
 serde_json.workspace = true
-gitbutler-storage.workspace = true
-but-secret.workspace = true
