+++ crates/but-forge-storage/src/storage.rs
@@ -9,13 +9,13 @@ const FORGE_SETTINGS_FILE: &str = "forge_settings.json";
 
 #[derive(Debug, Clone)]
 pub(crate) struct Storage {
-    inner: gitbutler_storage::Storage,
+    inner: but_fs::Storage,
 }
 
 impl Storage {
     pub fn from_path(path: impl Into<PathBuf>) -> Self {
         Storage {
-            inner: gitbutler_storage::Storage::new(path),
+            inner: but_fs::Storage::new(path),
         }
     }
