+++ crates/but-fs/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "gitbutler-fs"
+name = "but-fs"
 version = "0.0.0"
 edition = "2024"
 authors = ["GitButler <gitbutler@gitbutler.com>"]
@@ -9,9 +9,11 @@ rust-version = "1.89"
 [lib]
 doctest = false
 
+[features]
+legacy = []
+
 [dependencies]
 serde.workspace = true
-bstr.workspace = true
 anyhow.workspace = true
 gix = { workspace = true, features = ["dirwalk", "credentials", "parallel"] }
 walkdir = "2.5.0"
