+++ crates/but-github/Cargo.toml
@@ -14,11 +14,12 @@ rust-version = "1.89"
 doctest = false
 
 [dependencies]
-serde.workspace = true
 but-settings.workspace = true
 but-secret.workspace = true
 but-forge-storage.workspace = true
 but-error.workspace = true
+
+serde.workspace = true
 anyhow.workspace = true
 gitbutler-user.workspace = true
 serde_json.workspace = true
