+++ crates/gitbutler-forge/Cargo.toml
@@ -13,7 +13,7 @@ doctest = false
 serde.workspace = true
 serde_json = { workspace = true, features = ["arbitrary_precision"] }
 anyhow.workspace = true
-gitbutler-fs.workspace = true
+but-fs.workspace = true
 but-github.workspace = true
 git-url-parse = "0.6.0"
 but-forge-storage.workspace = true
