+++ crates/gitbutler-operating-modes/Cargo.toml
@@ -15,7 +15,7 @@ but-workspace = { workspace = true, features = ["legacy"] }
 but-core.workspace = true
 
 gitbutler-command-context.workspace = true
-gitbutler-fs.workspace = true
+but-fs.workspace = true
 gitbutler-stack.workspace = true
 
 serde.workspace = true
