+++ crates/gitbutler-oplog/Cargo.toml
@@ -19,7 +19,7 @@ but-oxidize.workspace = true
 gitbutler-command-context.workspace = true
 gitbutler-project.workspace = true
 gitbutler-branch.workspace = true
-gitbutler-fs.workspace = true
+but-fs.workspace = true
 gitbutler-reference.workspace = true
 gitbutler-stack.workspace = true
 gitbutler-repo.workspace = true
