+++ crates/gitbutler-project/Cargo.toml
@@ -14,8 +14,8 @@ but-error.workspace = true
 but-path.workspace = true
 but-serde =  { workspace = true, features = ["legacy"] }
 but-core.workspace = true
+but-fs = { workspace = true, features = ["legacy"] }
 
-gitbutler-storage.workspace = true
 gitbutler-forge.workspace = true
 
 anyhow.workspace = true
