+++ crates/gitbutler-stack/Cargo.toml
@@ -17,7 +17,7 @@ but-meta.workspace = true
 but-error.workspace = true
 but-oxidize.workspace = true
 
-gitbutler-fs.workspace = true
+but-fs.workspace = true
 gitbutler-command-context.workspace = true
 gitbutler-repo.workspace = true
 gitbutler-commit.workspace = true
