--- crates/gitbutler-storage/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "gitbutler-storage"
-version = "0.0.0"
-edition = "2024"
-authors = ["GitButler <gitbutler@gitbutler.com>"]
-publish = false
-rust-version = "1.89"
-
-[lib]
-doctest = false
-
-[dependencies]
-gitbutler-fs.workspace = true
