--- crates/gitbutler-storage/src/lib.rs
@@ -1,2 +1,0 @@
-mod storage;
-pub use storage::Storage;
