+++ crates/gitbutler-testsupport/Cargo.toml
@@ -15,6 +15,7 @@ but-settings.workspace = true
 but-oxidize.workspace = true
 but-meta = {workspace = true, features = ["legacy"]}
 but-workspace.workspace = true
+but-fs = {workspace = true, features = ["legacy"]}
 
 gitbutler-branch-actions = { path = "../gitbutler-branch-actions" }
 gitbutler-repo = { path = "../gitbutler-repo" }
@@ -22,7 +23,6 @@ gitbutler-command-context.workspace = true
 gitbutler-project.workspace = true
 gitbutler-user.workspace = true
 gitbutler-reference.workspace = true
-gitbutler-storage.workspace = true
 gitbutler-url.workspace = true
 gitbutler-stack.workspace = true
 gitbutler-commit.workspace = true
