+++ crates/gitbutler-user/Cargo.toml
@@ -12,8 +12,9 @@ doctest = false
 
 [dependencies]
 but-secret.workspace = true
-gitbutler-storage.workspace = true
+but-fs = { workspace = true, features = ["legacy"] }
 but-path.workspace = true
+
 anyhow.workspace = true
 serde.workspace = true
 serde_json = { workspace = true, features = ["arbitrary_precision"] }
